Pull build outer loop result up to abstract class
