+++ src/main/java/com/powsybl/openloadflow/ac/AcOuterLoopContext.java
@@ -20,8 +20,6 @@ public class AcOuterLoopContext extends AbstractOuterLoopContext<AcVariableType,
 
     private int iteration; // current iterations of this single outer loop type
 
-    private int outerLoopTotalIterations; // current total iterations over all outer loop types, for reporting purposes
-
     private AcSolverResult lastSolverResult;
 
     AcOuterLoopContext(LfNetwork network) {
@@ -45,11 +43,4 @@ public class AcOuterLoopContext extends AbstractOuterLoopContext<AcVariableType,
         this.lastSolverResult = lastSolverResult;
     }
 
-    public int getOuterLoopTotalIterations() {
-        return outerLoopTotalIterations;
-    }
-
-    public void setOuterLoopTotalIterations(int outerLoopTotalIterations) {
-        this.outerLoopTotalIterations = outerLoopTotalIterations;
-    }
 }
