+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/AbstractAreaInterchangeControlOuterLoop.java
@@ -10,6 +10,7 @@ import com.powsybl.openloadflow.network.LfBus;
 import com.powsybl.openloadflow.network.LfNetwork;
 import com.powsybl.openloadflow.network.util.ActivePowerDistribution;
 import com.powsybl.openloadflow.util.PerUnit;
+import com.powsybl.openloadflow.util.Reports;
 import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 
@@ -28,7 +29,7 @@ public abstract class AbstractAreaInterchangeControlOuterLoop<
         E extends Enum<E> & Quantity,
         P extends AbstractLoadFlowParameters<?>,
         C extends LoadFlowContext<V, E, P>,
-        O extends OuterLoopContext<V, E, P, C>>
+        O extends AbstractOuterLoopContext<V, E, P, C>>
         implements OuterLoop<V, E, P, C, O>,
         ActivePowerDistributionOuterLoop<V, E, P, C, O> {
 
@@ -138,7 +139,35 @@ public abstract class AbstractAreaInterchangeControlOuterLoop<
     }
 
     protected OuterLoopResult buildOuterLoopResult(Map<String, Pair<Set<LfBus>, Double>> areas, Map<String, ActivePowerDistribution.Result> resultByArea, ReportNode reportNode, O context) {
-        return null;
+        Map<String, Double> remainingMismatchByArea = resultByArea.entrySet().stream()
+                .filter(e -> !lessThanInterchangeMaxMismatch(e.getValue().remainingMismatch()))
+                .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().remainingMismatch()));
+        double totalDistributedActivePower = resultByArea.entrySet().stream().mapToDouble(e -> areas.get(e.getKey()).getRight() - e.getValue().remainingMismatch()).sum();
+        boolean movedBuses = resultByArea.values().stream().map(ActivePowerDistribution.Result::movedBuses).reduce(false, (a, b) -> a || b);
+        Map<String, Integer> iterationsByArea = resultByArea.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().iteration()));
+
+        ReportNode iterationReportNode = Reports.createOuterLoopIterationReporter(reportNode, context.getOuterLoopTotalIterations() + 1);
+        AreaInterchangeControlContextData contextData = (AreaInterchangeControlContextData) context.getData();
+        contextData.addDistributedActivePower(totalDistributedActivePower);
+        if (!remainingMismatchByArea.isEmpty()) {
+            logger.error(FAILED_TO_DISTRIBUTE_INTERCHANGE_ACTIVE_POWER_MISMATCH);
+            ReportNode failureReportNode = Reports.reportAreaInterchangeControlDistributionFailure(iterationReportNode);
+            remainingMismatchByArea.entrySet().stream().sorted(Map.Entry.comparingByKey()).forEach(entry -> {
+                logger.error("Remaining mismatch for Area {}: {} MW", entry.getKey(), entry.getValue() * PerUnit.SB);
+                Reports.reportAreaInterchangeControlAreaMismatch(failureReportNode, entry.getKey(), entry.getValue() * PerUnit.SB);
+            });
+            return handleDistributionFailure(context, contextData, movedBuses, totalDistributedActivePower, Double.NaN, FAILED_TO_DISTRIBUTE_INTERCHANGE_ACTIVE_POWER_MISMATCH);
+        } else {
+            if (movedBuses) {
+                areas.entrySet().stream().sorted(Map.Entry.comparingByKey()).forEach(entry -> {
+                    logger.info("Area {} interchange mismatch ({} MW) distributed in {} distribution iteration(s)", entry.getKey(), entry.getValue().getValue() * PerUnit.SB, iterationsByArea.get(entry.getKey()));
+                    Reports.reportAreaInterchangeControlAreaDistributionSuccess(iterationReportNode, entry.getKey(), entry.getValue().getValue() * PerUnit.SB, iterationsByArea.get(entry.getKey()));
+                });
+                return new OuterLoopResult(this, OuterLoopStatus.UNSTABLE);
+            } else {
+                return new OuterLoopResult(this, OuterLoopStatus.STABLE);
+            }
+        }
     }
 
     protected Set<LfBus> listBusesWithoutArea(LfNetwork network) {
