+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/AbstractOuterLoopContext.java
@@ -28,6 +28,7 @@ public abstract class AbstractOuterLoopContext<V extends Enum<V> & Quantity,
     protected Object data;
 
     protected C loadFlowContext;
+    private int outerLoopTotalIterations; // current total iterations over all outer loop types, for reporting purposes
 
     protected AbstractOuterLoopContext(LfNetwork network) {
         this.network = Objects.requireNonNull(network);
@@ -56,5 +57,13 @@ public abstract class AbstractOuterLoopContext<V extends Enum<V> & Quantity,
     @Override
     public void setLoadFlowContext(C loadFlowContext) {
         this.loadFlowContext = loadFlowContext;
+    }
+
+    public int getOuterLoopTotalIterations() {
+        return outerLoopTotalIterations;
+    }
+
+    public void setOuterLoopTotalIterations(int outerLoopTotalIterations) {
+        this.outerLoopTotalIterations = outerLoopTotalIterations;
     }
 }
